add yield algo

+++ src/split/esb/common.h
@@ -94,7 +94,7 @@ int service_init();
 uint32_t get_nonce();
 int process_payload(char* data, size_t length, uint32_t nonce);
 
-int handle_packet(struct zmk_split_esb_async_state* state);
+size_t handle_packet(struct zmk_split_esb_async_state* state);
 void reset_buffers();
